refactoring, add omdb logic, logger settings

+++ src/main/java/com/adamovichdev/movieawards/dao/entity/BaseEntity.java
@@ -12,13 +12,13 @@ public abstract class BaseEntity {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id")
-    protected int id;
+    protected Long id;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
